+++ Sources/Tokenizer.swift
@@ -2,7 +2,7 @@
 //  Tokenizer.swift
 //  SwiftFormat
 //
-//  Version 0.51.8
+//  Version 0.51.9
 //
 //  Created by Nick Lockwood on 11/08/2016.
 //  Copyright 2016 Nick Lockwood
@@ -47,7 +47,8 @@ private let swiftKeywords = Set([
     "fileprivate", "internal", "switch", "do", "catch", "enum", "struct", "throws",
     "throw", "typealias", "where", "break", "deinit", "subscript", "is", "while",
     "associatedtype", "inout", "continue", "operator", "repeat", "rethrows",
-    "default", "protocol", "defer", "await", /* Any, Self, self, super, nil, true, false */
+    "default", "protocol", "defer", "await", "consume",
+    /* Any, Self, self, super, nil, true, false */
 ])
 
 public extension String {
@@ -1494,7 +1495,12 @@ public func tokenize(_ source: String) -> [Token] {
                     lineStart = index(of: .nonSpaceOrComment, after: lineStart) ?? lineStart
                     switch tokens[lineStart] {
                     case .keyword("#elseif"), .keyword("#else"):
-                        lineStart = index(of: .startOfScope, before: lineStart) ?? lineStart
+                        while let start = index(of: .startOfScope, before: lineStart) {
+                            lineStart = start
+                            if tokens[start] == .startOfScope("#if") {
+                                break
+                            }
+                        }
                         fallthrough
                     case .startOfScope("#if"):
                         guard let prevIndex = index(of: .nonSpaceOrCommentOrLinebreak, before: lineStart) else {
@@ -1596,6 +1602,12 @@ public func tokenize(_ source: String) -> [Token] {
     func processToken() {
         var count = tokens.count
         var token = tokens[count - 1]
+        if !token.isSpaceOrComment, !token.isIdentifier,
+           let prevIndex = index(of: .nonSpaceOrComment, before: count - 1),
+           case .keyword("consume") = tokens[prevIndex]
+        {
+            tokens[prevIndex] = .identifier("consume")
+        }
         switch token {
         case let .keyword(name):
             if let prevIndex = index(of: .nonSpaceOrCommentOrLinebreak, before: count - 1),