+++ Tests/RulesTests+Wrapping.swift
@@ -280,6 +280,108 @@ class WrappingTests: RulesTests {
                        options: options)
     }
 
+    func testPreserveBlankLineBeforeElse() {
+        let input = """
+        if foo {
+            print("foo")
+        }
+
+        else if bar {
+            print("bar")
+        }
+
+        else {
+            print("baaz")
+        }
+        """
+
+        testFormatting(for: input, rule: FormatRules.elseOnSameLine)
+    }
+
+    func testPreserveBlankLineBeforeElseOnSameLine() {
+        let input = """
+        if foo {
+            print("foo")
+        }
+
+        else if bar {
+            print("bar")
+        }
+
+        else {
+            print("baaz")
+        }
+        """
+
+        let options = FormatOptions(elseOnNextLine: false)
+        testFormatting(for: input, rule: FormatRules.elseOnSameLine, options: options)
+    }
+
+    func testPreserveBlankLineBeforeElseWithComments() {
+        let input = """
+        if foo {
+            print("foo")
+        }
+        // Comment before else if
+        else if bar {
+            print("bar")
+        }
+
+        // Comment before else
+        else {
+            print("baaz")
+        }
+        """
+
+        testFormatting(for: input, rule: FormatRules.elseOnSameLine)
+    }
+
+    func testPreserveBlankLineBeforeElseDoesntAffectOtherCases() {
+        let input = """
+        if foo {
+            print("foo")
+        }
+        else {
+            print("bar")
+        }
+
+        guard foo else {
+            return
+        }
+
+        guard
+            let foo,
+            let bar,
+            lat baaz else
+        {
+            return
+        }
+        """
+
+        let output = """
+        if foo {
+            print("foo")
+        } else {
+            print("bar")
+        }
+
+        guard foo else {
+            return
+        }
+
+        guard
+            let foo,
+            let bar,
+            lat baaz
+        else {
+            return
+        }
+        """
+
+        let options = FormatOptions(elseOnNextLine: false, guardElsePosition: .nextLine)
+        testFormatting(for: input, output, rule: FormatRules.elseOnSameLine, options: options)
+    }
+
     // MARK: - wrapConditionalBodies
 
     func testGuardReturnWraps() {
@@ -2489,7 +2591,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 40)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_multipleTypealiases_beforeFirst() {
@@ -2512,7 +2614,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 45)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_afterFirst() {
@@ -2528,7 +2630,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 40)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_multipleTypealiases_afterFirst() {
@@ -2549,7 +2651,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 45)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth() {
@@ -2558,7 +2660,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 100)
-        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth_butWrappedInconsistently() {
@@ -2575,7 +2677,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 200)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth_butWrappedInconsistently2() {
@@ -2597,7 +2699,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 200)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth_butWrappedInconsistently3() {
@@ -2615,7 +2717,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 200)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth_butWrappedInconsistently4() {
@@ -2635,7 +2737,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .afterFirst, maxWidth: 200)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_shorterThanMaxWidth_butWrappedInconsistentlyWithComment() {
@@ -2655,7 +2757,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 200)
-        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_singleTypePreserved() {
@@ -2679,7 +2781,7 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 100)
-        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     func testWrapArguments_typealias_preservesCommentsAfterTypes() {
@@ -2691,7 +2793,24 @@ class WrappingTests: RulesTests {
         """
 
         let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 100)
-        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options)
+        testFormatting(for: input, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
+    }
+
+    func testWrapArguments_typealias_withAssociatedType() {
+        let input = """
+        typealias Collections = Collection<Int> & Collection<String> & Collection<Double> & Collection<Float>
+        """
+
+        let output = """
+        typealias Collections
+            = Collection<Int>
+            & Collection<String>
+            & Collection<Double>
+            & Collection<Float>
+        """
+
+        let options = FormatOptions(wrapTypealiases: .beforeFirst, maxWidth: 50)
+        testFormatting(for: input, output, rule: FormatRules.wrapArguments, options: options, exclude: ["sortTypealiases"])
     }
 
     // MARK: - -return wrap-if-multiline
@@ -4165,6 +4284,20 @@ class WrappingTests: RulesTests {
         """
         let options = FormatOptions(typeAttributes: .prevLine, fragment: true)
         testFormatting(for: input, rule: FormatRules.wrapAttributes, options: options)
+    }
+
+    func testClassAttributeNotMistakenForClassLet() {
+        let input = """
+        @objc final class MyClass: NSObject {}
+        let myClass = MyClass()
+        """
+        let output = """
+        @objc
+        final class MyClass: NSObject {}
+        let myClass = MyClass()
+        """
+        let options = FormatOptions(typeAttributes: .prevLine)
+        testFormatting(for: input, output, rule: FormatRules.wrapAttributes, options: options)
     }
 
     func testClassImportAttributeNotTreatedAsType() {